pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "piicloak"
version = "1.0.0"
description = "Enterprise-grade PII detection and anonymization API. GDPR/CCPA compliant. Supports 27+ entity types."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
keywords = ["pii", "pii-detection", "anonymization", "gdpr", "ccpa", "hipaa", "privacy", "data-protection", "presidio", "spacy", "nlp", "ner", "salesforce", "legal-tech", "fintech", "healthcare", "text-processing", "redaction", "compliance"]
authors = [
    {name = "Dmitry Marinov", email = "marinovdk@gmail.com"}
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Legal Industry",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Text Processing",
    "Environment :: Web Environment",
    "Framework :: Flask",
]
dependencies = [
    "flask>=3.0.0",
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
    "spacy>=3.7.0",
    "python-docx>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
production = [
    "gunicorn>=21.0.0",
]

[project.scripts]
piicloak = "piicloak.app:main"

[project.urls]
Homepage = "https://github.com/dimanjet/piicloak"
Documentation = "https://github.com/dimanjet/piicloak#readme"
Repository = "https://github.com/dimanjet/piicloak"
Issues = "https://github.com/dimanjet/piicloak/issues"
Changelog = "https://github.com/dimanjet/piicloak/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=piicloak --cov-report=term-missing"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
